+++ SanteDB.Persistence.MDM/Services/MdmResourceListener.cs
@@ -519,7 +519,7 @@ namespace SanteDB.Persistence.MDM.Services
         /// <param name="sourceEntity"></param>
         /// <param name="targetEntity"></param>
         /// <returns></returns>
-        private IdentifiedData CreateRelationship(Type relationshipType, Guid relationshipClassification, Guid? sourceEntity, Guid? targetEntity, Decimal? versionSequence)
+        private IdentifiedData CreateRelationship(Type relationshipType, Guid relationshipClassification, Guid? sourceEntity, Guid? targetEntity, Int32? versionSequence)
         {
             var relationship = Activator.CreateInstance(relationshipType, relationshipClassification, targetEntity) as IdentifiedData;
             relationship.Key = Guid.NewGuid();
